Correct the parameters to a function.
The name_redirect() function accepted two parameters, one for "name" and one for the rest of the path. This is
misleading for repositories that have a slash in the name (such as username/repo_name) because only the text before the
first slash will be called the "name" inside the function. This commit changes it to just take the entire path after
/v2/ and call that a "relative_path".