Improved logging during data load
crane previously logged only that it was about to start loading
metadata, and did not log when it completed.  Log both when it
completed, and (at debug level) log each file loaded.

This is important for us as we've had problems with delays
before data became available in crane, and without the log at
completion:

- we can't measure the performance of crane's data loading, so we can't
  judge if the performance is acceptable

- we can't know with certainty at what time published data has become
  available, so we can't effectively debug issues with missing data